Grand value division: value::from
